backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
sqlmodel==0.0.22
psycopg2-binary==2.9.9
alembic==1.13.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
